Create an Android Studio project and build files
Create an empty project using Android Studio, and edit it to create the
smallest possible project. Introduce a build script of Travis-CI
alongside this project.

There are a lot of files generated by Java, Android Studio, and other
tools that have no place in the repository. This is a "start as we mean
to go on" strategy. Using Travis makes it simplier to check
pull-requests before they're merged.