Bump the version of the framework used
Shift the commit of our Mother submodule from the tail of the previous
release to the tip of the final v1.0.0 release.

Numerous bug fixes, and missing functionality has been added to the
project since the last release. Mother's API is now stable.

Also incorporate data for Travis so that it is able to build the Mother
framework.