Create the entry point activity for Ash
Author a very bare "main activity" for Ash that inherits from the Mother
framework entry point. This activity overloads all state notifiers of
`android.app.Activity`.

The Mother framework requires a main activity to act as a single entry
point to control the behaviour of the different `Screen` objects, and
their layouts.

The only member of `MainActivity` immediately throws an
`UnsupportedOperationException` for self-explainitory reasons.
`MainActivity` is also registered in the manifest.