Extend `Screen` to create an initial "home screen"
Author the `HomeScreen` class that loads the 'activity_main.xml' layout
and presents it as the current content view. Make the required changes
to `MainActivity.getInitScreen` such that `HomeScreen` is loaded first.

Applications using the Mother framework create a series of "screens"
that move between each other to present different information to the
user.