Dynamically link together each of our screens
Register 'on click' listeners for each of the buttons in the
`home_screen` layout, and complete the `backButton()` overloads for each
screen.

`Screen`s are responsible for traversing through menus themselves. Each
screen is aware of which objects they should link to.

Also remove the misleading `@see` Javadoc from `SchoolsScreen` that is
clearly a copy-and-paste issue.