Update our manifest to allow us to use Google Maps
Include our Google Maps API key as metadata in our manifest, and also
state that our application requires access to the device's GPS.

The manifest file determines what our application can, and can not, do.
If we do not explicitly request these permissions we will be denied when
we formally request during execution.