Correct construction of local `services` variable
When constructing a list of services inside `ServicesScreen.resume()`
make use of the `String.equalsIgnoreCase(String)` and use its return
to return either `1`, or `0`.

The `compareTo` method normally returns a value of `-1`, `0`, or `1`
based on relativity. The overload found in the Mother framework does
not follow this system.

This fixes an issue where the local `services` variable becomes `null`
and propagates and exception when Android attempts to use it.