Add a title to the tag placed on our Google Map
Pass through the title of the tag when creating a new `LocationScreen`.

Currently our Google Map simply places a marker on itself when it is
created; it presents no indication about what that tag is. It would be
helpful to the user to inform them about what the marker represents.